LPD-28609 Move
