+++ modules/apps/asset/asset-publisher-web/src/test/java/com/liferay/asset/publisher/web/internal/display/context/AssetPublisherDisplayContextTest.java
@@ -8,8 +8,8 @@ package com.liferay.asset.publisher.web.internal.display.context;
 import com.liferay.asset.kernel.model.ClassType;
 import com.liferay.asset.kernel.model.ClassTypeReader;
 import com.liferay.asset.publisher.util.AssetPublisherHelper;
-import com.liferay.asset.publisher.web.internal.TestClassType;
 import com.liferay.asset.publisher.web.internal.configuration.AssetPublisherPortletInstanceConfiguration;
+import com.liferay.asset.publisher.web.internal.model.TestClassType;
 import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.portal.configuration.module.configuration.ConfigurationProviderUtil;
 import com.liferay.portal.kernel.module.configuration.ConfigurationException;
